+++ cyoa/app/assets/javascripts/users.js.coffee
