+++ cyoa/app/mailers/user_notification_mailer.rb
@@ -1,0 +1,2 @@
+class UserNotificationMailer < ApplicationMailer
+end
