+++ cyoa/app/views/layouts/mailer.html.erb
@@ -1,0 +1,5 @@
+<html>
+  <body>
+    <%= yield %>
+  </body>
+</html>
