+++ cyoa/app/views/layouts/mailer.text.erb
@@ -1,0 +1,1 @@
+<%= yield %>
