+++ cyoa/spec/mailers/user_notification_mailer_spec.rb
@@ -1,0 +1,5 @@
+require "rails_helper"
+
+RSpec.describe UserNotificationMailer, :type => :mailer do
+  pending "add some examples to (or delete) #{__FILE__}"
+end
